fix(app): mermaid rendering (#123)
Defers the rendering of graphs in `.mermaid` elements until they are
displayed. This ensures that the graphs are rendered properly with the
correct sizes.

I was able to reliably reproduce the issue by delaying the resolving of
the promises that flip the `hide` class on the `#loading` and `#content`
elements. The `.mermaid` graph element is part of the `#content` element
and initially this element is hidden. Depending on the load times the
`.mermaid` element may still be hidden which I assume caused the issues
of the rendering. The fix is to defer the initial default rendering via
`mermaid.initialize({ startOnLoad: false })` and manually trigger a
rendering via `mermaid.run()` once the `#content` is not hidden anymore,
which is the case once the promises are resolved.

---

Additionally I've added a
[doctype](https://developer.mozilla.org/en-US/docs/Glossary/Doctype) and
[html
language](https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Global_attributes/lang)
to the `index.html`.

Closes https://github.com/kumahq/kuma-counter-demo/issues/109

Signed-off-by: schogges <24444046+schogges@users.noreply.github.com>